horizontal rule conversion to ts
