+++ src/components/horizontal_rule/horizontal_rule.tsx
@@ -1,6 +1,19 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { SFC, HTMLAttributes } from 'react';
 import classNames from 'classnames';
+
+import { CommonProps } from '../common';
+
+export type EuiHorizontalRuleSize = 'full' | 'half' | 'quarter';
+
+export type EuiHorizontalRuleMargin = 'none' | 'xs' | 's' | 'm' | 'l' | 'xl' | 'xxl';
+
+export interface EuiHorizontalRuleProps {
+  /**
+   * Defines the width of the HR.
+   */
+  size?: EuiHorizontalRuleSize;
+  margin?: EuiHorizontalRuleMargin;
+}
 
 const sizeToClassNameMap = {
   full: 'euiHorizontalRule--full',
@@ -22,16 +35,18 @@ const marginToClassNameMap = {
 
 export const MARGINS = Object.keys(marginToClassNameMap);
 
-export const EuiHorizontalRule = ({
+export const EuiHorizontalRule: SFC<
+CommonProps & HTMLAttributes<HTMLHRElement> & EuiHorizontalRuleProps
+> = ({
   className,
-  size,
-  margin,
+  size = 'full',
+  margin = 'l',
   ...rest
 }) => {
   const classes = classNames(
     'euiHorizontalRule',
-    sizeToClassNameMap[size],
-    marginToClassNameMap[margin],
+    size ? sizeToClassNameMap[size] : undefined,
+    margin ? marginToClassNameMap[margin] : undefined,
     className
   );
 
@@ -41,13 +56,6 @@ export const EuiHorizontalRule = ({
       {...rest}
     />
   );
-};
-
-EuiHorizontalRule.propTypes = {
-  children: PropTypes.node,
-  className: PropTypes.string,
-  size: PropTypes.oneOf(SIZES),
-  margin: PropTypes.oneOf(MARGINS),
 };
 
 EuiHorizontalRule.defaultProps = {
