+++ src/components/horizontal_rule/index.ts
