+++ src/components/index.d.ts
@@ -14,7 +14,6 @@
 /// <reference path="./flyout/index.d.ts" />
 /// <reference path="./form/index.d.ts" />
 /// <reference path="./health/index.d.ts" />
-/// <reference path="./horizontal_rule/index.d.ts" />
 /// <reference path="./icon/index.d.ts" />
 /// <reference path="./key_pad_menu/index.d.ts" />
 /// <reference path="./link/index.d.ts" />
